requirements.txt
ftfy
alfred-py
